acpi: erst: cast size to uint64_t to clean up static analysis warning
CoverityScan warns of a possible integer overflow (which is very
unlikely to happen) when multiplying 32 bit integers.  Cast the
first to the same type of variable total_length to clear up this
warning.

Detected by CoverityScan, CID#1385375 ("Unintentional integer overflow")

Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>

+++ src/acpi/erst/erst.c
@@ -68,7 +68,7 @@ static int erst_test1(fwts_framework *fw)
 			table->length);
 		goto done;
 	}
-	total_length = (sizeof(fwts_acpi_serialization_instruction_entries) *
+	total_length = ((uint64_t)sizeof(fwts_acpi_serialization_instruction_entries) *
 		erst->instruction_entry_count) + sizeof(fwts_acpi_table_erst);
 	if (total_length > table->length) {
 		passed = false;
